rename panel
