+++ chat-app/src/components/Panel/index.js
